Add obscure text functionality
Create hiding and unhiding password input functionality
* Also seperated functions to seperate files
