+++ lib/Screens/Login/components/text_field_container.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+import '../../../constants.dart';
+
+class TextFieldContainer extends StatelessWidget {
+  final Widget child;
+  const TextFieldContainer({
+    Key key,
+    @required this.child,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    Size size = MediaQuery.of(context).size;
+    return Container(
+      margin: EdgeInsets.symmetric(vertical: 10),
+      padding: EdgeInsets.symmetric(vertical: 5, horizontal: 20),
+      decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(30), color: kPrimaryLightColor),
+      width: size.width * 0.8,
+      child: child,
+    );
+  }
+}
