+++ lib/Screens/Welcome/components/rounded_button.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 
-import '../../../constants.dart';
 
 class RoundedButton extends StatelessWidget {
   final String text;
